拾取范围变宽
